requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
firebirdsql==1.4.1
mysqlclient==2.2.7
PyJWT==2.10.1
sqlparse==0.5.3
pdfplumber==0.11.8
pdf2image==1.17.0
pytesseract==0.3.13
pillow==12.0.0
openai==2.8.1
python-dotenv==1.2.1
django-cors-headers==4.6.0
openpyxl==3.1.5
